requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
psycopg2-binary>=2.9,<3.0
pytest>=6.2,<7.0
pytest-django>=4.4,<5.0
djangorestframework-simplejwt>=4.7,<5.0